feat: display parent directory heading in localSearch results
