+++ src/client/theme-default/support/sidebar.ts
@@ -1,43 +1,9 @@
-import type { DefaultTheme } from 'vitepress/theme'
-import { ensureStartingSlash } from './utils'
-import { isActive } from '../../shared'
+import { isActive, type SidebarItem } from '../../shared'
 
 export interface SidebarLink {
   text: string
   link: string
   docFooterText?: string
-}
-
-type SidebarItem = DefaultTheme.SidebarItem
-
-/**
- * Get the `Sidebar` from sidebar option. This method will ensure to get correct
- * sidebar config from `MultiSideBarConfig` with various path combinations such
- * as matching `guide/` and `/guide/`. If no matching config was found, it will
- * return empty array.
- */
-export function getSidebar(
-  _sidebar: DefaultTheme.Sidebar | undefined,
-  path: string
-): SidebarItem[] {
-  if (Array.isArray(_sidebar)) return addBase(_sidebar)
-  if (_sidebar == null) return []
-
-  path = ensureStartingSlash(path)
-
-  const dir = Object.keys(_sidebar)
-    .sort((a, b) => {
-      return b.split('/').length - a.split('/').length
-    })
-    .find((dir) => {
-      // make sure the multi sidebar key starts with slash too
-      return path.startsWith(ensureStartingSlash(dir))
-    })
-
-  const sidebar = dir ? _sidebar[dir] : []
-  return Array.isArray(sidebar)
-    ? addBase(sidebar)
-    : addBase(sidebar.items, sidebar.base)
 }
 
 /**
@@ -106,14 +72,4 @@ export function hasActiveLink(
     : items.items
       ? hasActiveLink(path, items.items)
       : false
-}
-
-function addBase(items: SidebarItem[], _base?: string): SidebarItem[] {
-  return [...items].map((_item) => {
-    const item = { ..._item }
-    const base = item.base || _base
-    if (base && item.link) item.link = base + item.link
-    if (item.items) item.items = addBase(item.items, base)
-    return item
-  })
 }
